+++ knapsackkbest/src/kp_alg_data.c
@@ -8,8 +8,8 @@
 #include <stdlib.h>
 #include "../include/kp_alg/kp_alg.h"
 
-void kp_init_inn_sol(InnerSolution* iSol, uint16 n, uint16 j, uint16 t,
-uint16 v) {
+void kp_init_inn_sol(InnerSolution* iSol, uint32 n, uint32 j, uint32 t,
+uint32 v) {
 	*iSol = (InnerSolution) malloc(sizeof(struct innerSolution));
 	(*iSol)->dimension = n;
 	(*iSol)->column_idx = j;
@@ -17,7 +17,7 @@ uint16 v) {
 	(*iSol)->value = v;
 	(*iSol)->recovered = false;
 	// Sets the value to 0
-	(*iSol)->sol_vector = (uint16*) calloc(n, sizeof(uint16));
+	(*iSol)->sol_vector = (uint32*) calloc(n, sizeof(uint32));
 }
 
 void kp_free_inn_sol(InnerSolution innerSol) {
@@ -29,7 +29,7 @@ void kp_copy_inn_sol(InnerSolution* dest, InnerSolution origin) {
 	kp_init_inn_sol(dest, origin->dimension, origin->column_idx,
 			origin->row_idx, origin->value);
 	(*dest)->recovered = origin->recovered;
-	uint16 i;
+	uint32 i;
 	for (i = 0; i < origin->dimension; i++) {
 		(*dest)->sol_vector[i] = origin->sol_vector[i];
 	}
